Remove code redundancies
