+++ scm-plugins/scm-git-plugin/src/test/java/sonia/scm/repository/spi/GitModifyCommand_withEmptyRepositoryTest.java
@@ -24,27 +24,12 @@
 
 package sonia.scm.repository.spi;
 
-import com.github.sdorra.shiro.ShiroRule;
-import com.github.sdorra.shiro.SubjectAware;
 import org.eclipse.jgit.api.Git;
 import org.eclipse.jgit.api.errors.GitAPIException;
-import org.eclipse.jgit.errors.CorruptObjectException;
-import org.eclipse.jgit.lib.GpgSigner;
-import org.eclipse.jgit.lib.ObjectId;
-import org.eclipse.jgit.lib.ObjectReader;
 import org.eclipse.jgit.lib.Ref;
 import org.eclipse.jgit.revwalk.RevCommit;
-import org.eclipse.jgit.revwalk.RevWalk;
-import org.eclipse.jgit.treewalk.CanonicalTreeParser;
-import org.junit.BeforeClass;
-import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.TemporaryFolder;
-import sonia.scm.repository.GitTestHelper;
 import sonia.scm.repository.Person;
-import sonia.scm.repository.work.NoneCachingWorkingCopyPool;
-import sonia.scm.repository.work.WorkdirProvider;
-import sonia.scm.web.lfs.LfsBlobStoreFactory;
 
 import java.io.File;
 import java.io.IOException;
@@ -52,25 +37,8 @@ import java.nio.file.Files;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.Mockito.mock;
-import static sonia.scm.repository.spi.GitRepositoryConfigStoreProviderTestUtil.createGitRepositoryConfigStoreProvider;
 
-@SubjectAware(configuration = "classpath:sonia/scm/configuration/shiro.ini", username = "admin", password = "secret")
-public class GitModifyCommand_withEmptyRepositoryTest extends AbstractGitCommandTestBase {
-
-  @Rule
-  public TemporaryFolder temporaryFolder = new TemporaryFolder();
-  @Rule
-  public BindTransportProtocolRule transportProtocolRule = new BindTransportProtocolRule();
-  @Rule
-  public ShiroRule shiro = new ShiroRule();
-
-  private final LfsBlobStoreFactory lfsBlobStoreFactory = mock(LfsBlobStoreFactory.class);
-
-  @BeforeClass
-  public static void setSigner() {
-    GpgSigner.setDefault(new GitTestHelper.SimpleGpgSigner());
-  }
+public class GitModifyCommand_withEmptyRepositoryTest extends GitModifyCommandTestBase {
 
   @Test
   public void shouldCreateNewFileInEmptyRepository() throws IOException, GitAPIException {
@@ -133,20 +101,8 @@ public class GitModifyCommand_withEmptyRepositoryTest extends AbstractGitCommand
     return "sonia/scm/repository/spi/scm-git-empty-repo.zip";
   }
 
-  private void assertInTree(TreeAssertions assertions) throws IOException, GitAPIException {
-    try (Git git = new Git(createContext().open())) {
-      RevCommit lastCommit = getLastCommit(git);
-      try (RevWalk walk = new RevWalk(git.getRepository())) {
-        RevCommit commit = walk.parseCommit(lastCommit);
-        ObjectId treeId = commit.getTree().getId();
-        try (ObjectReader reader = git.getRepository().newObjectReader()) {
-          assertions.checkAssertions(new CanonicalTreeParser(null, reader, treeId));
-        }
-      }
-    }
-  }
-
-  private RevCommit getLastCommit(Git git) throws GitAPIException, IOException {
+  @Override
+  RevCommit getLastCommit(Git git) throws GitAPIException, IOException {
     return git.log().setMaxCount(1).all().call().iterator().next();
   }
 
@@ -162,18 +118,5 @@ public class GitModifyCommand_withEmptyRepositoryTest extends AbstractGitCommand
     request.addRequest(new ModifyCommandRequest.CreateFileRequest("new_file", newFile, false));
     request.setAuthor(new Person("Dirk Gently", "dirk@holistic.det"));
     return request;
-  }
-
-  private GitModifyCommand createCommand() {
-    return new GitModifyCommand(
-      createContext(),
-      new SimpleGitWorkingCopyFactory(new NoneCachingWorkingCopyPool(new WorkdirProvider())),
-      lfsBlobStoreFactory,
-      createGitRepositoryConfigStoreProvider());
-  }
-
-  @FunctionalInterface
-  private interface TreeAssertions {
-    void checkAssertions(CanonicalTreeParser treeParser) throws CorruptObjectException;
   }
 }
